+++ lib/screens/main_screen/main_screen.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: unnecessary_const
+
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
@@ -8,14 +10,30 @@ import '../../utils/image_assets.dart';
 import '../../utils/launch_url.dart';
 import '../../utils/themes.dart';
 import '../../widgets/custom_scaffold.dart';
+import 'widget/custom_card.dart';
+import 'package:date_picker_timeline/date_picker_timeline.dart';
 
 class MainScreen extends StatelessWidget {
-  const MainScreen({Key? key}) : super(key: key);
+  MainScreen({Key? key}) : super(key: key);
 
+  final colorList = [Colors.red, Colors.teal, Colors.blue];
+  final opacityList = const [
+    Color.fromRGBO(255, 0, 0, 0.2),
+    Color.fromARGB(51, 0, 255, 225),
+    Color.fromARGB(51, 0, 153, 255),
+  ];
   static const _sidePadding = EdgeInsets.symmetric(horizontal: 15);
-
   @override
   Widget build(BuildContext context) {
+    final _size = MediaQuery.of(context).size;
+    var _theme = Theme.of(context);
+    var _boxshadow = BoxShadow(
+      color: _theme.primaryColorDark,
+      spreadRadius: 2,
+      blurRadius: 8,
+      offset: const Offset(0, 1),
+    );
+
     return CustomScaffold(
       body: SafeArea(
         child: CustomScrollView(
@@ -34,41 +52,57 @@ class MainScreen extends StatelessWidget {
             ),
             SliverPadding(
               padding: const EdgeInsets.all(20),
-              sliver: SliverGrid.count(
-                crossAxisCount: 2,
-                childAspectRatio: 173 / 224,
-                crossAxisSpacing: 20,
-                mainAxisSpacing: 20,
-                children: const [
-                  DeptWidget(
-                    image: "aids",
-                    department: DepartmentEnum.aids,
-                  ),
-                  DeptWidget(
-                    image: "extc",
-                    department: DepartmentEnum.extc,
-                  ),
-                  DeptWidget(
-                    image: "cs",
-                    department: DepartmentEnum.cs,
-                  ),
-                  DeptWidget(
-                    image: "it",
-                    department: DepartmentEnum.it,
-                  ),
-                  DeptWidget(
-                    image: "biomed",
-                    department: DepartmentEnum.biomed,
-                  ),
-                  DeptWidget(
-                    image: "biotech",
-                    department: DepartmentEnum.biotech,
-                  ),
-                  DeptWidget(
-                    image: "chem",
-                    department: DepartmentEnum.chem,
-                  ),
-                ],
+              sliver: SliverList(
+                delegate: SliverChildListDelegate(
+                  [
+                    Container(
+                      width: _size.width * 0.9,
+                      decoration: BoxDecoration(
+                        color: _theme.primaryColor,
+                        borderRadius: BorderRadius.circular(15.0),
+                        border: Border.all(
+                          color: _theme.primaryColorLight,
+                          width: 1,
+                          style: BorderStyle.solid,
+                        ),
+                        boxShadow: [_boxshadow],
+                      ),
+                      child: DatePicker(
+                        DateTime.now(),
+                        monthTextStyle: _theme.textTheme.subtitle2!,
+                        dayTextStyle: _theme.textTheme.subtitle2!,
+                        dateTextStyle: _theme.textTheme.subtitle2!,
+                        initialSelectedDate: DateTime.now(),
+                        selectionColor: Colors.blue,
+                        daysCount: 7,
+                      ),
+                    ),
+                    const SizedBox(
+                      height: 20.0,
+                    ),
+                    Padding(
+                      padding: const EdgeInsets.symmetric(
+                        horizontal: 4.0,
+                        vertical: 15.0,
+                      ),
+                      child: SizedBox(
+                        height: _size.height * 0.6,
+                        width: _size.width,
+                        child: ListView.builder(
+                            physics: const BouncingScrollPhysics(),
+                            itemCount: 3,
+                            itemBuilder: (context, value) {
+                              var color = colorList[value];
+                              var opacity = opacityList[value];
+                              return CustomCard(
+                                color,
+                                opacity,
+                              );
+                            }),
+                      ),
+                    ),
+                  ],
+                ),
               ),
             )
           ],
@@ -76,52 +110,59 @@ class MainScreen extends StatelessWidget {
       ),
     );
   }
+
+  //
+
+  // @override
+  // Widget build(BuildContext context) {
+
+  // }
 }
 
-class DeptWidget extends StatelessWidget {
-  const DeptWidget({
-    Key? key,
-    required this.image,
-    required this.department,
-  }) : super(key: key);
+// class DeptWidget extends StatelessWidget {
+//   const DeptWidget({
+//     Key? key,
+//     required this.image,
+//     required this.department,
+//   }) : super(key: key);
 
-  final String image;
-  final DepartmentEnum department;
+//   final String image;
+//   final DepartmentEnum department;
 
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: () => GoRouter.of(context).push(
-        "/department?department=${department.index}",
-      ),
-      child: Card(
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(12),
-        ),
-        margin: EdgeInsets.zero,
-        color: Theme.of(context).colorScheme.secondary,
-        child: Padding(
-          padding: const EdgeInsets.all(15),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Flexible(
-                child: Padding(
-                  padding: const EdgeInsets.all(20.0),
-                  child: Image.asset(
-                    "assets/images/branches/$image.png",
-                    height: 100,
-                  ),
-                ),
-              ),
-              Text(department.name),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
+//   @override
+//   Widget build(BuildContext context) {
+//     return GestureDetector(
+//       onTap: () => GoRouter.of(context).push(
+//         "/department?department=${department.index}",
+//       ),
+//       child: Card(
+//         shape: RoundedRectangleBorder(
+//           borderRadius: BorderRadius.circular(12),
+//         ),
+//         margin: EdgeInsets.zero,
+//         color: Theme.of(context).colorScheme.secondary,
+//         child: Padding(
+//           padding: const EdgeInsets.all(15),
+//           child: Column(
+//             mainAxisAlignment: MainAxisAlignment.spaceBetween,
+//             children: [
+//               Flexible(
+//                 child: Padding(
+//                   padding: const EdgeInsets.all(20.0),
+//                   child: Image.asset(
+//                     "assets/images/branches/$image.png",
+//                     height: 100,
+//                   ),
+//                 ),
+//               ),
+//               Text(department.name),
+//             ],
+//           ),
+//         ),
+//       ),
+//     );
+//   }
+// }
 
 class MainScreenAppBar extends ConsumerWidget {
   const MainScreenAppBar({
